+++ wpcleanadmin/includes/ajax/user-roles-ajax.php
@@ -3,8 +3,8 @@
  * WPCleanAdmin User Roles AJAX Class
  *
  * @package WPCleanAdmin
- * @version 1.8.4
- * @author Sut
+ * @version 1.8.5
+ * @author Tanox
  * @author URI: https://github.com/Tanox
  * @since 1.7.15
  */