+++ wpcleanadmin/includes/class-wpca-extension-api.php
@@ -3,8 +3,8 @@
  * WPCleanAdmin Extension API Class
  *
  * @package WPCleanAdmin
- * @version 1.8.4
- * @author Sut
+ * @version 1.8.5
+ * @author Tanox
  * @author URI: https://github.com/Tanox
  * @since 1.8.0
  * @description Extension API for third-party developers to extend WP Clean Admin functionality