+++ wpcleanadmin/includes/modules/admin/settings/assets/js/wpca-settings.js
@@ -2,8 +2,8 @@
  * WP Clean Admin Settings JavaScript
  * 
  * @package WPCleanAdmin\Modules\Admin\Settings
- * @version 1.8.4
- * @author Sut
+ * @version 1.8.5
+ * @author Tanox
  * @since 1.8.0
  */