+++ wpcleanadmin/includes/settings/fields/class-wpca-diagnostics-settings-fields.php
@@ -3,8 +3,8 @@
  * WPCleanAdmin Diagnostics Settings Fields Class
  *
  * @package WPCleanAdmin
- * @version 1.8.4
- * @author Sut
+ * @version 1.8.5
+ * @author Tanox
  * @author URI: https://github.com/Tanox
  * @since 1.8.0
  */