+++ wpcleanadmin/wp-clean-admin.php
@@ -3,8 +3,8 @@
  * Plugin Name: WP Clean Admin
  * Plugin URI: https://github.com/Tanox/WP-Clean-Admin
  * Description: A comprehensive WordPress admin cleanup and optimization plugin
- * Version:  1.8.4
- * Author: Sut
+ * Version: 1.8.4
+ * Author: Tanox
  * Author URI: https://github.com/Tanox
  * License: GPL v2 or later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html
@@ -77,7 +77,7 @@ if ( ! function_exists( '\plugin_basename' ) ) {
 }
 
 // Define plugin constants
-define( 'WPCA_VERSION', '1.8.4' );
+define( 'WPCA_VERSION', '1.8.5' );
 define( 'WPCA_PLUGIN_DIR', ( \function_exists( '\plugin_dir_path' ) ? \plugin_dir_path( __FILE__ ) : \dirname( __FILE__ ) . '/' ) );
 define( 'WPCA_PLUGIN_URL', ( \function_exists( '\plugin_dir_url' ) ? \plugin_dir_url( __FILE__ ) : '' ) );
 define( 'WPCA_TEXT_DOMAIN', 'wp-clean-admin' );